chore: moved styles from index to styled
